Add hideReadLinks and showAllLinks fx
Adds hideReadLinks which targets and hides all read links on click of the
'Show Unread Button'. Function is behaving as desired. Uncovered bug which
would prevent user from alternating between show read and show unread. Added
showAllLinks function which is invoked prior to execution of hideUnreadLinks
and hideReadLinks functions. This enables the user to go back and forth between
read and unread links with no need to refresh.